dumbledore/backend/requirements.txt
fastapi
uvicorn
python-multipart
openai            # if using OpenAI API
openai-whisper    # optional for local
pyttsx3
